% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jamiedata.R
\name{ds_cleaner}
\alias{ds_cleaner}
\title{Clean DS}
\usage{
ds_cleaner(data, dif, noise)
}
\arguments{
\item{data}{csv file with the PMIS data}

\item{dif}{upper percentile threshold used to eliminate spikes}

\item{noise}{percentile used to determine the level of sensor noise}
}
\description{
Clean DS
}
